fix(displays): preserve wake ownership across reloads

+++ crates/dormant-displays/src/macos_power.rs
@@ -34,6 +34,7 @@ use core_foundation_sys::base::kCFAllocatorDefault;
 use core_foundation_sys::string::{CFStringCreateWithCString, CFStringRef, kCFStringEncodingUTF8};
 use dormant_core::error::E_DISPLAY_IO;
 use dormant_core::types::CmdFailure;
+use libc::c_int;
 
 use crate::macos_display_sleep::{
     AssertionGuard, CGDirectDisplayID, DisplaySleepTransport, PMSET_SLEEPNOW_ARGS,
@@ -104,9 +105,9 @@ unsafe extern "C" {
         displayCount: *mut u32,
     ) -> CgError;
 
-    /// `Boolean CGDisplayIsAsleep(CGDirectDisplayID display)` — Apple's
-    /// `Boolean` is `unsigned char`; nonzero means asleep.
-    fn CGDisplayIsAsleep(display: CGDirectDisplayID) -> u8;
+    /// SAFETY: `CGDisplayConfiguration.h:288` declares this as returning
+    /// Mach `boolean_t`, which is C `int`; nonzero means asleep.
+    fn CGDisplayIsAsleep(display: CGDirectDisplayID) -> c_int;
 }
 
 /// Build a `CFStringRef` from a Rust `&str` via `CFStringCreateWithCString`
